Refactored to move menu related code to menu.js
