+++ public/menu.js
@@ -7,7 +7,8 @@
  *
  *  Licensed under the MIT License. See LICENSE in the repository root for license information.
  *--------------------------------------------------------------------------------------------*/
-const { Menu, BrowserWindow, app, shell } = require("electron");
+const { Menu, BrowserWindow, app, shell, Tray, MenuItem } = require("electron");
+const path = require("path");
 
 let devToolsAccelerator = "Cmd+Option+I";
 let alwaysOnTopMenuItem;
@@ -15,6 +16,14 @@ let copyScreenshotMenuItem;
 let saveScreenshotMenuItem;
 let startRecordingMenuItem;
 let stopRecordingMenuItem;
+
+// Tray-related variables
+let tray = null;
+let trayContextMenu = null;
+let trayStartRecordingItem = null;
+let trayStopRecordingItem = null;
+
+const isMacOS = process.platform === "darwin";
 
 function createMenu(mainWindow, displayWindow, packageInfo) {
   const template = [
@@ -237,9 +246,316 @@ function updateRecordingMenuItems(displayVisible, isRecording) {
   }
 }
 
+// Tray functionality
+function createTray(mainWindow, displayWindow, packageInfo) {
+  if (!isMacOS) return null;
+
+  // Create tray icon - use the dedicated menuicon.png
+  let trayIconPath = path.join(__dirname, "../images/menuicon.png");
+
+  tray = new Tray(trayIconPath);
+  // Try to set template mode if the method exists
+  if (typeof tray.setTemplate === "function") {
+    tray.setTemplate(true);
+  }
+
+  createTrayMenu(mainWindow, displayWindow, packageInfo);
+  tray.setToolTip("Carabiner - Screen Capture and Remote Control");
+
+  // Click to show context menu (no window toggling)
+  tray.on("click", () => {
+    tray.popUpContextMenu();
+  });
+
+  return tray;
+}
+
+function createTrayMenu(
+  mainWindow,
+  displayWindow,
+  packageInfo,
+  captureDevices = null,
+  settings = null,
+  isCurrentlyRecording = false
+) {
+  const trayMenu = Menu.buildFromTemplate([
+    {
+      label: "Show Carabiner",
+      click: () => {
+        // Only show the display window (screen capture window)
+        if (displayWindow && !displayWindow.isVisible()) {
+          displayWindow.show();
+        }
+      },
+    },
+    { type: "separator" },
+    {
+      label: "Start Recording",
+      id: "tray-start-recording",
+      accelerator: "CmdOrCtrl+Shift+R",
+      enabled: false,
+      click: () => {
+        if (displayWindow) {
+          if (!displayWindow.isVisible()) {
+            displayWindow.show();
+          }
+          displayWindow.webContents.send("start-recording");
+        }
+      },
+    },
+    {
+      label: "Stop Recording",
+      id: "tray-stop-recording",
+      accelerator: "CmdOrCtrl+Shift+S",
+      enabled: false,
+      click: () => {
+        if (displayWindow) {
+          displayWindow.webContents.send("stop-recording");
+        }
+      },
+    },
+  ]);
+
+  appendCaptureDevicesMenu(trayMenu, mainWindow, captureDevices, settings);
+  trayMenu.append(new MenuItem({ type: "separator" }));
+  trayMenu.append(
+    new MenuItem({
+      label: "Settings...",
+      accelerator: "CmdOrCtrl+,",
+      click: () => {
+        mainWindow.webContents.send("open-display-tab");
+        mainWindow.show();
+      },
+    })
+  );
+  trayMenu.append(new MenuItem({ type: "separator" }));
+
+  trayMenu.append(
+    new MenuItem({
+      label: "Documentation",
+      accelerator: "F1",
+      click: () => {
+        shell.openExternal(`${packageInfo.repository.url}#readme`);
+      },
+    })
+  );
+  trayMenu.append(
+    new MenuItem({
+      label: "Keyboard Control",
+      accelerator: "CmdOrCtrl+F1",
+      click: () => {
+        shell.openExternal(`${packageInfo.repository.url}/blob/main/docs/key-mappings.md`);
+      },
+    })
+  );
+  trayMenu.append(
+    new MenuItem({
+      label: "Release Notes",
+      click: () => {
+        shell.openExternal(`${packageInfo.repository.url}/releases`);
+      },
+    })
+  );
+  trayMenu.append(new MenuItem({ type: "separator" }));
+  trayMenu.append(new MenuItem({ role: "quit" }));
+
+  // Set the tray context menu
+  trayContextMenu = trayMenu;
+  // Store references to the recording menu items for later updates
+  trayStartRecordingItem = trayContextMenu.getMenuItemById("tray-start-recording");
+  trayStopRecordingItem = trayContextMenu.getMenuItemById("tray-stop-recording");
+  updateTrayRecordingMenuItems(isCurrentlyRecording);
+  // Set the context menu for the tray
+  if (tray) {
+    tray.setContextMenu(trayContextMenu);
+  }
+}
+
+function appendCaptureDevicesMenu(menu, mainWindow, captureDevices = null, settings = null) {
+  if (!captureDevices || captureDevices.length === 0) {
+    // If no capture devices are available, just return
+    return;
+  }
+  menu.append(new MenuItem({ type: "separator" }));
+  captureDevices.forEach((device) => {
+    let deviceLabel = device.label || `Device ${captureDevices.indexOf(device) + 1}`;
+    const found = settings?.control?.deviceList?.find((d) => d.linked === device.deviceId);
+    deviceLabel += found ? ` - ${found.type} ${found.alias ?? found.ipAddress}` : "";
+    menu.append(
+      new MenuItem({
+        label: deviceLabel,
+        type: "radio",
+        checked: settings?.display?.deviceId === device.deviceId,
+        click: () => {
+          mainWindow.webContents.send("update-capture-device", device.deviceId);
+        },
+      })
+    );
+  });
+}
+
+function updateTrayRecordingMenuItems(isRecording) {
+  if (!tray || !trayStartRecordingItem || !trayStopRecordingItem) return;
+
+  // Start recording is always enabled when not recording (it can show the window if needed)
+  trayStartRecordingItem.enabled = !isRecording;
+
+  // Stop recording is only enabled when currently recording
+  trayStopRecordingItem.enabled = isRecording;
+}
+
+function toggleDockIcon(showInDock, mainWindow, displayWindow = null, packageInfo = null) {
+  if (!isMacOS) return null;
+
+  // Ensure we have window references
+  const display =
+    displayWindow ||
+    BrowserWindow.getAllWindows().find((win) => win.webContents.getURL().includes("display.html"));
+
+  if (showInDock) {
+    app.dock.show();
+    if (tray) {
+      tray.destroy();
+      tray = null;
+      trayContextMenu = null;
+      trayStartRecordingItem = null;
+      trayStopRecordingItem = null;
+    }
+  } else {
+    // Create tray first
+    if (!tray && mainWindow && display && packageInfo) {
+      createTray(mainWindow, display, packageInfo);
+    }
+
+    // Hide dock icon
+    app.dock.hide();
+  }
+
+  return tray;
+}
+
+function createContextMenu(
+  mainWindow,
+  displayWindow,
+  packageInfo,
+  isCurrentlyRecording,
+  captureDevices = null,
+  settings = null
+) {
+  const menu = new Menu();
+  let fullscreenAcc = "F11";
+  if (isMacOS) {
+    fullscreenAcc = "Cmd+Ctrl+F";
+  }
+
+  menu.append(
+    new MenuItem({
+      label: "Copy Screenshot",
+      accelerator: "CmdOrCtrl+Shift+C",
+      click: () => {
+        displayWindow.webContents.send("copy-screenshot");
+      },
+    })
+  );
+  menu.append(
+    new MenuItem({
+      label: "Save Screenshot As...",
+      accelerator: "CmdOrCtrl+S",
+      click: () => {
+        displayWindow.webContents.send("save-screenshot");
+      },
+    })
+  );
+  menu.append(new MenuItem({ type: "separator" }));
+  menu.append(
+    new MenuItem({
+      id: "start-recording-ctx",
+      label: "Start Recording",
+      accelerator: "CmdOrCtrl+Shift+R",
+      enabled: !isCurrentlyRecording,
+      click: () => {
+        displayWindow.webContents.send("start-recording");
+      },
+    })
+  );
+  menu.append(
+    new MenuItem({
+      id: "stop-recording-ctx",
+      label: "Stop Recording",
+      accelerator: "CmdOrCtrl+Shift+S",
+      enabled: isCurrentlyRecording,
+      click: () => {
+        displayWindow.webContents.send("stop-recording");
+      },
+    })
+  );
+  menu.append(new MenuItem({ type: "separator" }));
+  menu.append(
+    new MenuItem({
+      label: "Paste Text",
+      accelerator: "CmdOrCtrl+V",
+      click: () => {
+        displayWindow.webContents.send("handle-paste");
+      },
+    })
+  );
+  menu.append(new MenuItem({ type: "separator" }));
+  menu.append(
+    new MenuItem({
+      role: "togglefullscreen",
+      accelerator: fullscreenAcc,
+    })
+  );
+  menu.append(
+    new MenuItem({
+      label: "Hide Screen",
+      click: () => {
+        displayWindow.hide();
+      },
+    })
+  );
+  appendCaptureDevicesMenu(menu, mainWindow, captureDevices, settings);
+  menu.append(new MenuItem({ type: "separator" }));
+  menu.append(
+    new MenuItem({
+      label: "Settings...",
+      accelerator: "CmdOrCtrl+,",
+      click: () => {
+        mainWindow.webContents.send("open-display-tab");
+        mainWindow.show();
+      },
+    })
+  );
+  menu.append(new MenuItem({ type: "separator" }));
+  menu.append(
+    new MenuItem({
+      label: "Keyboard Control Help",
+      accelerator: "CmdOrCtrl+F1",
+      click: () => {
+        shell.openExternal(`${packageInfo.repository.url}/blob/main/docs/key-mappings.md`);
+      },
+    })
+  );
+  menu.append(new MenuItem({ type: "separator" }));
+  menu.append(new MenuItem({ role: "quit" }));
+
+  return menu;
+}
+
+// Getter functions for tray state
+function getTray() {
+  return tray;
+}
+
 module.exports = {
   createMenu,
   updateAlwaysOnTopMenuItem,
   updateScreenshotMenuItems,
   updateRecordingMenuItems,
+  createTrayMenu,
+  appendCaptureDevicesMenu,
+  updateTrayRecordingMenuItems,
+  toggleDockIcon,
+  createContextMenu,
+  getTray,
 };
